% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbindCIs.R
\name{rbindCIs}
\alias{rbindCIs}
\title{Merge results from profileCI function}
\usage{
rbindCIs(x)
}
\arguments{
\item{x}{List. A list of \code{closedp} objects to be merged into a single \code{matrix}.}
}
\value{
A \code{matrix} where each row represents model outputs and columns are variables,
suchs as number of individuals, intervals...
}
\description{
To summarize results more efficients, use this function to merge them into one object.
}
\author{
Roman Lustrik \email{roman.lustrik@biolitika.si}
}

