\name{Rcapture-package}
\alias{Rcapture-package}
\alias{Rcapture}
\docType{package}
\title{ Loglinear Models in Capture-Recapture Experiments }
\description{
   Estimation of abundance and other demographic parameters for closed populations, open populations and the robust design in capture-recapture experiments using loglinear models.
}
\details{
\tabular{ll}{
Package: \tab Rcapture\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2008-01-28\cr
License: \tab GPL\cr
}
}
\author{ Sophie Baillargeon \email{Sophie.Baillargeon@mat.ulaval.ca} and\cr
Louis-Paul Rivest \email{Louis-Paul.Rivest@mat.ulaval.ca} }
\references{
Baillargeon, S. and Rivest, L.P. (2007). Rcapture: Loglinear models for capture-recapture in R.  \emph{Journal of Statistical Software}, \bold{19}(5), http://www.jstatsoft.org/

Chao, A. (1987). Estimating the population size for capture-recapture data with unequal catchability. \emph{Biometrics}, \bold{43}(4), 783--791.

Cormack, R. M. (1985). Example of the use of glim to analyze capture-recapture studies. In \emph{Lecture Notes in Statistics 29: Statistics in Ornithology}, Morgan, B. et North, P. editors, Springer-Verlag, New York, 242--274.

Cormack, R. M. (1989). Log-linear models for capture-recapture. \emph{Biometrics}, \bold{45}, 395--413.

Cormack, R. M. (1992). Interval estimation for mark-recapture studies of closed populations. \emph{Biometrics}, \bold{48}, 567--576.

Cormack, R. M. (1993). Variances of mark-recapture estimates. \emph{Biometrics}, \bold{49}, 1188--1193.

Cormack, R. M. and Jupp, P. E. (1991). Inference for Poisson and multinomial models for capture-recapture experiments. \emph{Biometrika}, \bold{78}(4), 911--916.

Rivest, L.P. and Levesque, T. (2001). Improved log-linear model estimators of abundance in capture-recapture experiments. \emph{Canadian Journal of Statistics}, \bold{29}, 555--572.

Rivest, L.P. and Daigle, G. (2004). Loglinear models for the robust design in mark-recapture experiments. \emph{Biometrics}, \bold{60}, 100--107.

Rivest, L.P. and Baillargeon, S. (2007). Applications and extensions of Chao's moment estimator for the size of a closed population. \emph{Biometrics}, \bold{63}(4), 999--1006.

Rivest, L.P. (2008). Why a time effect often has a limited impact on capture-recapture estimates in closed populations. \emph{Canadian Journal of Statistics}, \bold{36}(1), to be published. 
}
\keyword{ package }
